feat: Update routes page to use RoutesView component
